refactor(ng-dev/pr): move fetch-pull-request into common directory
As part of #203, refactoring the pr directory into more reusable pieces.
