+++ ng-dev/pr/merge/pull-request.ts
@@ -4,19 +4,13 @@
  *
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
- */
-
-import {params, types as graphqlTypes} from 'typed-graphqlify';
-
-import {parseCommitMessage} from '../../commit-message/parse';
-import {AuthenticatedGitClient} from '../../utils/git/authenticated-git-client';
-import {getPr} from '../../utils/github';
+ */ import {parseCommitMessage} from '../../commit-message/parse';
 import {PullRequestFailure} from './failures';
 import {matchesPattern} from './string-pattern';
 import {PullRequestMergeTask} from './task';
 import {getTargetBranchesForPullRequest} from './target-label';
 import {assertCorrectBreakingChangeLabeling, assertPendingState} from './validations';
-import {fetchPullRequestFromGithub} from './fetch-pull-request';
+import {fetchPullRequestFromGithub} from '../common/fetch-pull-request';
 
 /** Interface that describes a pull request. */
 export interface PullRequest {
